projects/invoicepipes/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "invoice-extractor"
dynamic = ["version"]
description = "Simple CLI tool for extracting data from invoice documents"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "quantalogic-py-zerox<=0.0.8",
    "pydantic>=2.5.0",
    "python-magic>=0.4.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.1.0",
]

[project.scripts]
invoice-extractor = "invoice_extractor.cli:main"

[tool.hatch.version]
path = "src/invoice_extractor/__init__.py"

[tool.ruff]
line-length = 88
target-version = "py311"
